+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.2.3] - 2023-10-22
+
+### Fixed
+
+- [[#51](https://github.com/kislerdm/terraform-provider-neon/issues/51)] Fixed credentials content.
+- Fixed management of the `resource_role` state:
+  - Fixed password reading.
+  - Removed the side effect upon the resource import: the role's password won't be reset now. 
+
+### Changed
+
+- Updated dependencies: Neon Go SDK [v0.2.5](https://github.com/kislerdm/neon-sdk-go/releases/tag/v0.2.5)
+
 ## [v0.2.2] - 2023-09-19
 
 ### Fixed